AI_model/requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
PyPDF2>=3.0.0
python-docx>=0.8.11
requests>=2.28.0
python-linkedin-v2>=0.8.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
scikit-learn>=1.2.0
tqdm>=4.65.0
logging>=0.5.1.2 